Vlab Issue #56329.
Updates to CHGRES so it can work with GFS data older than 2007.

Fix logic that checks for the existance of an input NSST file.
When not processing an NSST file, the script would set the
NSST file name to NULL.  The working directory will
sometimes contain a zero byte NULL file, which CHGRES
thinks is an actual NSST file.

All changes are backwards compatible.

Fixes #56329.

Change-Id: I55dda23fbd0bd99dae538237bde71cbb903bfc52