Mirror updates from workflow and cmakeify (#3)
Updates all the programs with changes that were made in global workflow and adds a build script in order to build the utilities. As part of this, the machine-setup.sh script is copied from workflow to the ush directory.

Also adds cmake capability to most of the remaining programs. The programs that used to live under `util` in workflow were not previously converted to cmake.

The rdbfmsua program is still not converted. There are complications there due to the reliance on gfortran-specific string manipulation functions even though it is build with intel.