TST: Write tests for finddate.sh
I wrote tests for most of the functions in finddate.sh in order.  I suppose the main remaining one is a check of the results of the examples in the file header against the results from the old version.  That'll take care of the date/sequence parsing in the finddate function as well.

I used the newest version of TAP so I could group tests within a single file.  Output should be fairly intuitive even without a TAP parser on hand.