Add ocean and ice products to SFS (#117)
This PR adds ocean_sfs.csv and ice_sfs.csv files for the ocnicepost.fd
utility. It also addresses a bug in the utils_mod.F90 program of the
utility, where the vector pairs loop was crashing when there was more
than one pair with the same dimensions and remapping method. Resolves
Issue #113 .