Fix ice temperatures in sfcanl file when running with CICE (#159)
# Description

When creating sfcanl files in the GFSv17 framework with CICE, the fields
tsoil1/2/3/4 get populated with constant values of 200K or 271.2K over
ice grids. This results in bad initialization when used in downstream
applications like GEFS. The replacement should only be done when ice is
present and these constant values exist. The use of these constant
values as a check also preserves functionality when not using CICE.

Resolves #158